Converted the storage of a container from int to float
Attached the tracker goal amount to the variable

Added variable to store the weekly and monthly amount

+++ app/src/main/java/com/example/mwt/db/dateprogressdb/DateProgressEntity.kt
@@ -6,7 +6,7 @@ import androidx.room.PrimaryKey
 @Entity(tableName = "dateProgress")
 data class DateProgressEntity(
         val date: String,
-        val progress: Int
+        val progress: Float
 ) {
     @PrimaryKey(autoGenerate = true)
     var id: Long = 0
